Moved notifications form to notifications page
